Possibility add a paragraph that's a translation of another paragraph.
Adding r=tr to paragraph properties (must also have rd and rp/ra) allows
to reference to a paragraph, that the new paragraph will be a translation
of. If the original paragraph is changed, it reflects to the readers of
the translated versions too as a yellow bar.

Translating areas is not yet supported and will cause wonky results.