Templates: After creating a new document a user can load content specified in an existing document located in document folder/Templates or any of its parent folders (assuming the user has manage rights).
Create a copy button:
Added a copy document button to Manage page. Now a user can create a copy of a document. It should be useful with Templates.

Edit settings:
Added Edit settings button in cogwheel menu. This is a shortcut to edit attr settings paragraph(s) or to create a new one.