No more duplicated area references in edit window
DocParagraph.get_exported_markdown returns text from the source
paragraph if there's none specified in a translation. This used
to have a side effect of duplicating area references in the edit
window. This commit disables the behavior for area references and
adds some tests to make sure everything looks sane.