Add a list for whitelisted external URLs for downloading images
Add a whitelist to the module, where one can configure hostnames to trust.
If the filter encounters an image with an url that points to an external host,
it checks the whitelist to see if it is allowed to download the image and then
attach it to the output LaTeX document.