Fetch velp data in the same way as in the old promise-based version
This is faster because more routes are called in parallel, but it's not immediately clear if all of them can be called in parallel, so to be on the safe side, let's not do that and just imitate what the old version did.

Make also sure the default velp group won't be added twice to the collection.