Add URL routes for confirming and expiring view rights
The use case is to have these URLs in tableform cells (generated by jsrunner), so they are quickly accessible, without having to go to manage page and search for the user.

These sort of URLs are not supposed to be a long-term solution. Ideally, tableform should support Angular components in cells and then these confirm/expire buttons would be components.