Fix lecture wall display logic
Until now, the visibility of the lecture wall was dependent on the most-recently-active lecture tab. So, if the user turned off the wall on one tab and then activated another tab where the wall was visible, then a newly-opened tab would have the wall visible, even though the user's latest action was to turn off the wall.

With this change, the newly-opened tab will obey the user's latest action instead of the most-recently-active tab.