Add config option INHERIT_FOLDER_RIGHTS_DOCS
This enables inheriting document rights from its containing folder.
This could possibly be a document setting in the future.

We don't use the blockassociation table for this because it's currently
intended for uploaded files and images that don't exist in the TIM folder hierarchy.
For example, if you moved a document to some other folder,
the association would still point to the old folder.

Some day the uploaded files could be visible in the folder hierarchy,
and then it would make sense to unify the two mechanisms.