Add environment_classes texmacro, allow emitting environments for some CSS classes
Adds a texenvironment_classes under texmacros. With the option, it becomes possible to mark some CSS classes to be handled as LaTeX environments. This is needed for paragraph classes for some of them to support verbatim environments.

The option allows for soft conversion of paragraph classes into LaTeX environments without breaking existing printing. In the