calendar: Fix group filter priority
If a user belonged to a booker group g1 and there was an event for groups g1, g2, if a calendar filter was set to only show g2's events, it would still show the event for the user. This was because the filter for group names came after filter for all events the user can see via eventgroups.

This fix moves the group filter to ensure that the filter would leave out these discrepancies.