Fix /view auto redirect not working for locked duration rights
Before when a duration right was active but not yet unlocked, going to /teacher (or other non-view route) would not auto redirect because no active permission was found. This fix adjusts the check by allowing to search for unused duration permissions as well. If one is found, user is redirected to /view and is shown the common "you're about to unlock the document"