group: Add ability to send welcome message to new group users
This feature allows to set an automatic welcome message to be sent to the users that are added to a user group. To specify the message, save it as a global answer to the user group management document with task name "GLO_join_message".
The contents of this global answer are used as the message contents.