settings: Add support for quick themes
Quick themes are toggle buttons in the Settings tab that allow to quickly enable and disable user themes. The selections are preserved on a per-session basis, allowing to modify which styles to keep enabled for different devices.

For example, the feature can be used to add a toggle button for one's favorite dark theme and have it easily be enabled or disabled for different devices.