upload: add /docUploads route
The route allows to list all files uploaded into a document as a JSON list. This can be used by automation tools to list available file and their paths. In the future, the path can also be used by TIM itself to build a GUI for listing all files in a document.