docinfo: add extraGroupPreambleFolder
The `extraGroupPreambleFolder` preamble document setting allows loading preambles based on what groups a user belongs to.

The setting takes a TIM path and only functions within a preamble. When a document is loaded with a preamble that includes the option, the following occurs:

1. The path in `extraGroupPreambleFolder` is searched for any documents with the short name of form `preamble-<group_name>-<priority>` where `<group_name>` is a name of a user group that the current user loading the page is a member of and `<priority>` is an integer.
2. The documents with the highest priority are loaded as preambles. If multiple documents have the same priority, they all are loaded in lexicographical order.

The group preambles are loaded in an inverse order; that is, the document with the highest priority is loaded last and it thus overrides the settings of the other group preambles.

The normal preamble loading order still applies; the most local preamble is always loaded last.