scim: add SCIM_ALLOW_UPDATE_HAKA_USER_INFO option
The option allows for a temporary workaround the cases where a user has previously logged in via HAKA but their username is not the same as what SCIM provides. This may be the case for users outside the SCIM's organization. Since the user's username is needed on many cases (e.g. grade export), it maybe reasonable to change the name.

This commit also adds a FIXME regarding implementing a more robust option, where the user's username is preserved but the SCIM username is saved in a way that the grade exporter can use.